requirements.txt
# Core dependencies
python-dotenv>=1.0.0
openai>=1.3.0

# Flask & WebSockets
flask>=3.0.0
flask-socketio>=5.0.0
python-socketio>=5.0.0
python-engineio>=4.0.0
werkzeug>=3.0.0

# Database
psycopg[binary]>=3.1.0
psycopg-pool>=3.1.0

# LangGraph & LangChain
langgraph>=0.0.20
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10

# Data processing
numpy>=1.24.0
pandas>=2.0.0
pydantic>=2.0.0

# Jupyter (for client)
jupyter>=1.0.0
jupyterlab>=4.0.0
ipywidgets>=8.0.0
ipywidgets==8.1.1

# Development
black==23.12.0
flake8==6.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
